requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
boto3==1.34.113
botocore==1.34.113
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
openai==1.30.3
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.2
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
google-cloud-texttospeech
